+++ frontend/src/public/components/TemplateEdit/ExtraFields/constants.ts
@@ -1,0 +1,3 @@
+import { EExtraFieldType } from '../../../types/template';
+
+export const DATASET_FIELD_TYPES = [EExtraFieldType.Checkbox, EExtraFieldType.Radio, EExtraFieldType.Creatable];
